feat : 인기글 등록 로직 좋아요 10개 기준 컬럼은 글쓸때 선택 인기글 전체 조회, 컬럼 전체 조회

+++ src/main/java/com/BitOracle/BitOracle/dto/PortfolioResDto.java
@@ -12,7 +12,7 @@ public class PortfolioResDto {
     private String coinName;
     private BigDecimal quantity;
     private  double averageBuyPrice;
-    private BigDecimal totalBuyAmount;
+    private double totalBuyAmount;
     private double currentPrice;
     private double evaluatuonAmount;
     private double profitRate;
